Add the ability to create addresses to the RedeemAll method
This change also renames the method to ReedeemAllAndCreateAddresses. The reason
for the complexity of the commit is that we need to have a single top-level
contract call that will cause many transactions to occur in the same block. Some
of those transactions should cause accounts to be deleted, and others should
cause accounts to be created.

This way, we can tickle a bug in the current implementation of validation which
occurs when account deletions and additions take place out of order and the
validation logic isn't able to find a preimage for a particular account.