Add a BigMap contract for testing the storge tried
This contracts makes it easy to add a bunch of storage slots in a single
transation, and then in a separate transaction to free a bunch of those slots
and add some others.

The idea is that this sort of operation will pass with the deterministic flag on
the storage db implementation set and fail when it is not set.

Part of NIT-3068