Multi redeem public (#313)
* Add a method for redeeming a list of retryable ticket ids

This will allow us to test that when many retryables are redeemed in the same
block, that the execution client doesn't have any problems.

Related to: NIT-3010

* Add the gas cost to the call

* Add the ability to create addresses to the RedeemAll method

This change also renames the method to ReedeemAllAndCreateAddresses. The reason
for the complexity of the commit is that we need to have a single top-level
contract call that will cause many transactions to occur in the same block. Some
of those transactions should cause accounts to be deleted, and others should
cause accounts to be created.

This way, we can tickle a bug in the current implementation of validation which
occurs when account deletions and additions take place out of order and the
validation logic isn't able to find a preimage for a particular account.

* Convert tabs to spaces

* format: yarn format

---------

Co-authored-by: gzeon <im@gzeon.dev>
Co-authored-by: gzeon <hng@offchainlabs.com>