Add MintNativeToken and BurnNativeToken calls
These will allow arbOwner authorized addresses to mint some amount of the
chain's native gas token to or burn some amount of the chain's native gas token
from a given address.

Chain owners *should* deploy a non-upgradable proxy contract, the address of
which is added to the set of arbOwner addresses and then add any addresses which
should be allowed to mint and burn the native gas token to call corresponding
methods on the proxy contract. The implmentation of those proxy contract methods
should not allow to and from address arguments, and instead use the address of
the caller as the target for minting and burning.

Resolves: NIT-3297